requirements.txt
python-dotenv>=1.0.0
langchain>=0.1.16
langchain-community>=0.0.21
chromadb>=0.4.24
tavily>=0.3.2
ollama>=0.1.9
